Refactored python code
